Fixed the silly gcc 4.0 warning about ~0 being negative 1.
git-svn-id: file:///fltk/svn/fltk/branches/branch-1.1@4356 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ test/colbrowser.cxx
@@ -193,7 +193,7 @@ search_entry(int r, int g, int b)
     int i, j, diffr, diffg, diffb;
     unsigned int diff, mindiff;
 
-    mindiff = ~0;
+    mindiff = (unsigned int)~0;
     for (i = j = 0; db->r < 256; db++, i++)
     {
        diffr = r - db->r;
